+++ tests/unit_tests/execution/test_execution_engine.py
@@ -36,6 +36,7 @@ from nautilus_trader.model.events import AccountState
 from nautilus_trader.model.events import OrderCancelled
 from nautilus_trader.model.events import OrderUpdated
 from nautilus_trader.model.identifiers import AccountId
+from nautilus_trader.model.identifiers import ClientId
 from nautilus_trader.model.identifiers import ClientOrderId
 from nautilus_trader.model.identifiers import InstrumentId
 from nautilus_trader.model.identifiers import PositionId
@@ -102,7 +103,7 @@ class ExecutionEngineTests(unittest.TestCase):
 
         self.venue = Venue("SIM")
         self.exec_client = MockExecutionClient(
-            name=self.venue.value,
+            client_id=ClientId(self.venue.value),
             account_id=self.account_id,
             engine=self.exec_engine,
             clock=self.clock,
@@ -117,7 +118,7 @@ class ExecutionEngineTests(unittest.TestCase):
         result = self.exec_engine.registered_clients
 
         # Assert
-        self.assertEqual(["SIM"], result)
+        self.assertEqual([ClientId("SIM")], result)
 
     def test_deregister_client_removes_client(self):
         # Arrange
@@ -278,6 +279,7 @@ class ExecutionEngineTests(unittest.TestCase):
     def test_given_random_command_logs_and_continues(self):
         # Arrange
         random = TradingCommand(
+            AUDUSD_SIM.id.venue.client_id,
             self.trader_id,
             self.account_id,
             AUDUSD_SIM.id,
@@ -316,7 +318,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -354,7 +356,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -391,7 +393,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -427,7 +429,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -487,7 +489,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_bracket = SubmitBracketOrder(
-            entry.instrument_id,
+            entry.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -547,7 +549,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_bracket1 = SubmitBracketOrder(
-            entry1.instrument_id,
+            entry1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -573,7 +575,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_bracket2 = SubmitBracketOrder(
-            entry2.instrument_id,
+            entry2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -636,7 +638,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_bracket1 = SubmitBracketOrder(
-            entry1.instrument_id,
+            entry1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -662,7 +664,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_bracket2 = SubmitBracketOrder(
-            entry2.instrument_id,
+            entry2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -714,7 +716,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -751,7 +753,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -789,7 +791,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -853,7 +855,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -869,9 +871,10 @@ class ExecutionEngineTests(unittest.TestCase):
         self.exec_engine.process(TestStubs.event_order_filled(order, AUDUSD_SIM))
 
         cancel_order = CancelOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
+            order.instrument_id,
             order.client_order_id,
             VenueOrderId("1"),
             self.uuid_factory.generate(),
@@ -906,7 +909,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -922,9 +925,10 @@ class ExecutionEngineTests(unittest.TestCase):
         self.exec_engine.process(TestStubs.event_order_filled(order, AUDUSD_SIM))
 
         update_order = UpdateOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
+            order.instrument_id,
             order.client_order_id,
             Quantity(200000),
             order.price,
@@ -959,7 +963,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1010,7 +1014,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1059,7 +1063,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1110,7 +1114,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1174,7 +1178,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1238,7 +1242,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1296,7 +1300,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1370,7 +1374,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1439,7 +1443,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1457,7 +1461,7 @@ class ExecutionEngineTests(unittest.TestCase):
         expected_position_id = PositionId("P-19700101-000000-000-001-1")
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1518,7 +1522,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1538,7 +1542,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1615,7 +1619,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy1.id,
@@ -1626,7 +1630,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy2.id,
@@ -1736,7 +1740,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy1.id,
@@ -1749,7 +1753,7 @@ class ExecutionEngineTests(unittest.TestCase):
         position_id1 = PositionId("P-1")
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy1.id,
@@ -1760,7 +1764,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order3 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy2.id,
@@ -1853,7 +1857,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1873,7 +1877,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1946,7 +1950,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order1 = SubmitOrder(
-            order1.instrument_id,
+            order1.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -1966,7 +1970,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order2 = SubmitOrder(
-            order2.instrument_id,
+            order2.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,
@@ -2034,7 +2038,7 @@ class ExecutionEngineTests(unittest.TestCase):
         )
 
         submit_order = SubmitOrder(
-            order.instrument_id,
+            order.instrument_id.venue.client_id,
             self.trader_id,
             self.account_id,
             strategy.id,